preperation for nodelets
This diff applies changes to the structure on how the nodes work. The goal
is to have a structure that allows laterone to transform them into
nodelets.

To archive nodelet compatible structure, I have to remove *::run() methods and the
while(ros::ok()); loops.

The current run method has actually two parts:
  1. initialization
  2. the main loop

Because of this I decided to move the initialization part into a own
function. The main loop is put into a "periodicUpdate"
function, which is registerd with a ros timer for periodic call.

This is done for navsat_transformation_node and ukf/ekf_localiaztion_nodes.